refactor: extract canvas download to shared utility
DRY up the canvas download logic used in AlgorithmBitmap components
by extracting it to a reusable downloadCanvasAsPng utility function.

The download filename format is algorithmId_seed.png (e.g. myalgo_1-2-3.png).
