+++ apps/app/src/lib/utils.ts
@@ -1,7 +1,7 @@
 import { clsx } from 'clsx'
 import { twMerge } from 'tailwind-merge'
 import type { ClassValue } from 'clsx'
-import type { FamilyKind } from '@entropretty/utils'
+import type { AlgorithmId, FamilyKind } from '@entropretty/utils'
 
 export function cn(...inputs: Array<ClassValue>) {
   return twMerge(clsx(inputs))
@@ -27,4 +27,18 @@ export function familyKindLabel(familyKind: FamilyKind) {
     case 'ProceduralPersonal':
       return 'Personal Id'
   }
+}
+
+export function downloadCanvasAsPng(
+  canvas: HTMLCanvasElement,
+  algorithmId: AlgorithmId,
+  seed: Array<number>,
+) {
+  const link = document.createElement('a')
+  link.download = `${algorithmId}_${seed.join('-')}.png`
+  link.href = canvas.toDataURL('image/png')
+  link.click()
+  setTimeout(() => {
+    link.remove()
+  }, 0)
 }
